User switch not handled properly in the accessibility manager service.
On user switch the transient state of the old user was not cleared
which means that when we switch back to this user the operational
state such as which event types were dispatched, what state was sent
to local managers, etc is stale. This leads to semi-updated state
and broken behavior. Now if the user becomes inactive, we are clearing
all transient state which will be recreated when the user becomes
active.

bug:8196652

Change-Id: Ie9e0d712b6d567e5074b328f1bb87afaa5395c06