Reset reverse state when an animator ends
Seeking an animation after the animator has reverse()'d to the beginning
will result in seeking in reverse. Starting the animation will make it proceed
normally, and calling reverse() will also give expected behavior. But seeking
causes this unexpected behavior because the state of reversing is not reset until
the next time the animation is played (either by start() or reverse()).

Fix is to reset the internal flag when the animator ends.

Issue #8234676 Reversing an Animator Leaves it in A Reversed State Until Explicitly Started

Change-Id: I9d212ae1879aa277d1add7eb4c7ec61432af059e