Fix bug with transientState for listviews without stable IDs
This is the other part of a fix with transient state. A layout container
may be out of sync with its adapter during a data change. When the transient
state views are managed by their positions, and these positions may not match
between the container and the adapter, then the views will not be updated
correctly on the screen (though the underlying data will still be correct).

An earlier fix addressed the problem when the adapter has stable IDs (managing
the transient views by their IDs instead of their positions). This fix addresses
the other part of the problem, simply avoiding managing via positions during
a data change.

Issue #8254775 View.setHasTransient state has side-effects when deleting content in ListView

Change-Id: I0fe1692a1507a042180f8a14a8ef2d0c6946a061