Fix keyguard pattern invalidate bug
The framework used to invalidate both the previous invalidate
region and the current one. However, with change I9fb96f99
this is no longer the case (it was always true for software
rendering, but never showed because most platforms use
acceleration).

The fix is to invalidate the union of the old region and
the new region ourselves.

Fixes bug 8503031

Change-Id: If303e0892772dd9d6915efa55118d054a742c1a5