Eradicate basically all of the system reboot stuff in the watchdog.
This was implemented for 1.0 out of paranoia about the possibility
of needing to schedule reboots of the system after it went out in
the field, which I am happy to say was never ever used.

Let's get rid of it.

A small stub is left, to still service the API that allows people
to send a reboot broadcast to have the system immediately reboot.

Change-Id: I6731b24a28340e50c8015f8cb28e48f74f69f9b7